requirements.txt
Faker
redis
psycopg2-binary
pandas
plotly[express]